backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
dnspython==2.4.2